Improve `storage_alias` and make `UnlockAndUnreserveAllFunds` independent of the pallet (#14773)
* Make `storage_alias` more generic over the `prefix`

* Make `UnlockAndUnreserveAllFunds` indepenend from the pallet

* FMT

* Fix error reporting

* Rename prefix type

* Add test

* Apply suggestions from code review


* ".git/.scripts/commands/fmt/fmt.sh"

---------
